export types migrated and null conditionals implemented
