+++ web-interface/components/CommentContainer.tsx
@@ -22,7 +22,7 @@ const CommentContainer = ({commentsArray, justSent, arrowDown, color} : CommentC
         <>
             <div className={openedCard.commentsContainer}>
                 {commentsArray.map((element: any, index)=> 
-                    <Comment key={element.key} text={element.text} last_modified={element.last_modified} avatar={element.author.avatar} color={color}/>
+                    <Comment key={element.key} text={element.text} last_modified={element.last_modified} avatar={element.author.avatar} color={color} username={element.author.username}/>
                 )} 
             </div>
         </>
