+++ web-interface/components/InterfaceDropdown.tsx
@@ -3,65 +3,11 @@ import 'bootstrap/dist/css/bootstrap.min.css'
 import {Dropdown, DropdownItem, DropdownMenu, DropdownToggle} from 'reactstrap'
 import { useRouter } from 'next/router'
 import dashboard from '../styles/Dashboards.module.css';
-
-type parent_columns = {
-    parent_id: number,
-    parent_name: string,
-    parent_section: number,
-    parent_position: number,
-} 
-
-interface workSpace {
-    workspace_id: number,
-    type: number,
-    is_archived: number,
-    name: string,
-}
-
-type card = {
-    "card_id": number,
-    "custom_id": number | null,
-    "title": string,
-    "owner_user_id": number | null,
-    "owner_username": string | null,
-    "owner_avatar": string | null,
-    "type_id": number | null,
-    "color": string,
-    "section": number,
-    "lane_id": number,
-    "position": number
-  }
-  
-type column = {
-  "column_id": number,
-  "workflow_id": number,
-  "section": number,
-  "parent_column_id": Array<parent_columns> | null ,
-  "position": number,
-  "name": string,
-  "description": string,
-  "color": string,
-  "limit": number,
-  "cards_per_row": number,
-  "flow_type": number,
-  "card_ordering": string | null,
-  "cards": Array<card> | null,
-  "order": number
-}
-  
-  
-type workflow = {
-    "type": number,
-    "position": number,
-    "is_enabled": number,
-    "is_collapsible": number,
-    "name": string,
-    "workflow_id": number,
-    "columns": Array<column> | null
-}
+import { type } from 'os';
+import { workSpace, workflow } from '@/types/types';
 
 interface DropdownProps {
-    data : Array<workSpace> | Array<workflow>
+    data : Array<workSpace | workflow>
     getData : any
     name : string
 }
