+++ web-interface/pages/board/[board_id].tsx
@@ -10,52 +10,9 @@ import {urlCloud} from '../../constants'
 import dashboard from '../../styles/Dashboards.module.css';
 import Cookies from 'cookies';
 import {useRouter} from 'next/router';
-import {card} from '../../components/CardsWorkflow';
+import { workflow, card, } from '@/types/types';
 
 type Props = {}
-
-type user = {
-    user_id: number | null,
-    username: string,
-    realname: string,
-    avatar: string
-}
-
-type parent_columns = {
-    parent_id: number,
-    parent_name: string,
-    parent_section: number,
-    parent_position: number,
-}
-
-type column = {
-  "column_id": number,
-  "workflow_id": number,
-  "section": number,
-  "parent_column_id": Array<parent_columns> | null ,
-  "position": number,
-  "name": string,
-  "description": string,
-  "color": string,
-  "limit": number,
-  "cards_per_row": number,
-  "flow_type": number,
-  "card_ordering": string | null,
-  "cards": Array<card> ,
-  "order": number
-}
-
-
-type workflow = {
-  "type": number,
-  "position": number,
-  "is_enabled": number,
-  "is_collapsible": number,
-  "name": string,
-  "workflow_id": number,
-  "users": Array<user>,
-  "columns": Array<column>
-}
 
   type NextJsI18NConfig = {
     defaultLocale: string
@@ -70,7 +27,7 @@ type workflow = {
   }
 
   interface PropsResponse {
-    data : Array<workflow>
+    data : Array<workflow> 
     _nextI18Next : NextJsI18NConfig
   }
 
@@ -114,10 +71,19 @@ const Board = ( props: PropsResponse) => {
   }
 
   const moveCards = (current : number, cardIndex : number, destiny: number ) =>{
+    
     const tempWorkflow = workflow;
-    tempWorkflow.columns[destiny].cards.push(tempWorkflow.columns[current].cards[cardIndex]);
-    tempWorkflow.columns[current].cards.splice(cardIndex, 1);
-    setWorkflow(tempWorkflow)
+
+    if(tempWorkflow!=null){
+      
+      if(tempWorkflow.columns[current].cards != null){
+        tempWorkflow.columns[destiny].cards?.push(tempWorkflow.columns[current].cards[cardIndex]);
+      }
+      
+      tempWorkflow.columns[current].cards?.splice(cardIndex, 1);
+      setWorkflow(tempWorkflow)
+  
+    }
   }
 
   const showModal = (value: boolean) =>{
@@ -129,9 +95,7 @@ const Board = ( props: PropsResponse) => {
     <>
     
     <div className={dashboard.modalWrap}>
-        {/*displayCard && currentCard?.owner_avatar!= null && currentCard?.owner_username!=null  && <OpenedActivityCard title={currentCard.title} owner={currentCard.owner_username} owner_avatar={currentCard.owner_avatar} co_owner_usernames={currentCard.co_owner_usernames} co_owner_avatars={currentCard.co_owner_avatars} description={currentCard.description} setDisplayCard={setDisplayCard}/>*/}
-        {/*<OpenedActivityCard title={"T2.HU10.- Crear componente de sidebar con Perfil, organización, idioma, link a workspaces y cerrar sesión."} owner={""} owner_avatar={"https://s3.amazonaws.com/kanbamne/attachments/university6y/avatar_80x80_10.jpg"} co_owner_usernames={[]} co_owner_avatars={[]}  description={"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in "}/>*/}
-    
+
         {displayCard && currentCard!=undefined && <OpenedActivityCard title={currentCard.title} owner={currentCard.owner_username} owner_avatar={currentCard.owner_avatar} co_owner_usernames={currentCard.co_owner_usernames} co_owner_avatars={currentCard.co_owner_avatars} description={currentCard.description} setDisplayCard={setDisplayCard} color={currentCard.color} card_id={currentCard.card_id} comment_count={currentCard.comment_count}/>}
     
     </div>
